Refactoring plugin structure
