+++ src/plugins/index.ts
@@ -1,0 +1,1 @@
+export { logger } from './logger'
