+++ src/plugins/logger/config/index.ts
